+++ web/main/test/functional/fixtures/textblocks.json
@@ -34,5 +34,17 @@
     "content": "",
     "doc_class": "Text"
   }
+},
+{
+  "model": "main.textblock",
+  "pk": 4,
+  "fields": {
+    "created_at": "2023-02-21T15:36:18.701",
+    "updated_at": "2023-02-21T15:37:31.082",
+    "name": "notes",
+    "description": null,
+    "content": "<p>This is the first paragraph. This sentence is highlighted.</p><p>This is the second paragraph. This sentence is corected.</p><p>This is the third paragraph. This and the following paragraph are elided.</p><p>This paragraph is also elided.</p><p>This is the last paragraph. It's normal.</p>",
+    "doc_class": "Text"
+  }
 }
 ]
